WIP, better fixture scoping and cleanup

+++ test/functional/test_examples.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
+# -**-
 from __future__ import absolute_import
 from __future__ import print_function
 
@@ -44,7 +44,8 @@ class Example(object):
             assert k8s_obj is not None
 
     def test_patch_resource(self, ansible_helper, resources, patch_params, obj_compare):
-        resource = next(resources())
+        resource_gen = resources()
+        resource = next(resource_gen)
         for params in patch_params:
             name = params.get('name')
             namespace = params.get('namespace')
@@ -59,7 +60,8 @@ class Example(object):
             obj_compare(ansible_helper, new_obj, params)
 
     def test_replace_resource(self, ansible_helper, resources, replace_params, obj_compare):
-        resource = next(resources())
+        resource_gen = resources()
+        resource = next(resource_gen)
         for params in replace_params:
             name = params.get('name')
             namespace = params.get('namespace')
@@ -93,39 +95,44 @@ class Example(object):
     def patch_params(self, project, object_name):
         patch_tasks = filter(lambda x: x.get('patch'), self.tasks['tasks'])
         parameters = map(lambda x: x['patch'], patch_tasks)
-        for parameter in parameters:
+        for i, parameter in enumerate(parameters):
             if parameter.get('namespace'):
                 parameter['namespace'] = project
-            parameter['name'] = object_name
+            parameter['name'] = '{}-{}'.format(i, object_name)[:23].strip('-')
         return parameters
 
     @pytest.fixture()
     def replace_params(self, project, object_name):
         replace_tasks = filter(lambda x: x.get('replace'), self.tasks['tasks'])
         parameters = map(lambda x: x['replace'], replace_tasks)
-        for parameter in parameters:
+        for i, parameter in enumerate(parameters):
             if parameter.get('namespace'):
                 parameter['namespace'] = project
-            parameter['name'] = object_name
+            parameter['name'] = '{}-{}'.format(i, object_name)[:23].strip('-')
         return parameters
 
     @pytest.fixture()
     def resources(self, ansible_helper, create_params):
         def resource_generator():
             for create in create_params:
+                exception = None
                 request_body = ansible_helper.request_body_from_params(create)
                 namespace = create.get('namespace')
                 name = create.get('name')
                 k8s_obj = ansible_helper.create_object(namespace, body=request_body)
 
-                yield k8s_obj
-
                 try:
-                    ansible_helper.delete_object(name, namespace)
-                except KubernetesException as ex:
-                    # Swallow exception if object is already removed
-                    if ex.value.get('status') != 404:
-                        raise
+                    yield k8s_obj
+                except Exception as e:
+                    exception = e
+                finally:
+                    try:
+                        if not exception:
+                            ansible_helper.delete_object(name, namespace)
+                    except KubernetesException as ex:
+                        # Swallow exception if object is already removed
+                        if ex.value.get('status') != 404:
+                            raise
         return resource_generator
